+++ settings.php
@@ -1,0 +1,3 @@
+<?php
+require_once('functions.php');
+require_once('helpers.php');
